Simplify wrapping of iterator into KVIterator.
